+++ burn-tensor/src/tensor/api/int.rs
@@ -11,7 +11,7 @@ where
     ///
     /// * `range` - The range of values to generate.
     /// * `device` - The device to create the tensor on.
-    pub fn arange(range: Range<usize>, device: &B::Device) -> Self {
+    pub fn arange(range: Range<i64>, device: &B::Device) -> Self {
         Tensor::new(B::float_arange(range, device))
     }
 
@@ -21,7 +21,7 @@ where
     ///
     /// * `range` - The range of values to generate.
     /// * `step` - The step between each value.
-    pub fn arange_step(range: Range<usize>, step: usize, device: &B::Device) -> Self {
+    pub fn arange_step(range: Range<i64>, step: usize, device: &B::Device) -> Self {
         Tensor::new(B::float_arange_step(range, step, device))
     }
 }
